+++ baseframe/assets.py
@@ -85,6 +85,7 @@ assets['jquery.fullcalendar.css'][Version('1.6.4')] = ('jquery.ui.css', 'basefra
 assets['jquery.fullcalendar.gcal.js'][Version('1.6.4')] = ('jquery.fullcalendar.js', 'baseframe/js/fullcalendar/gcal.js')
 
 assets['jquery.locationpicker.js'][Version('0.1.12')] = ('jquery.js', 'baseframe/js/jquery.locationpicker.js')
+assets['jquery.appear.js'][Version('0.3.3')] = ('jquery.js', 'baseframe/js/jquery.appear.js')
 
 assets['select2.js'][Version('3.3.2')] = ('jquery.js>=1.4.6', 'baseframe/js/select2.js')
 assets['select2.css'][Version('3.3.2')] = 'baseframe/css/select2.css'